refactor chessboard into multiple smaller util classes
