+++ src/chess_insights/util/enum_file_and_rank.py
