+++ src/chess_insights/util/enum_square.py
