+++ tests/chess_pieces/test_chess_piece.py
