+++ tests/chess_pieces/test_white_king.py
