--- tests/run_tests_with_coverage.txt
@@ -1,5 +1,0 @@
-pytest -v --cov=./chess_insights --cov-report=html
-python -m pytest -v --cov=./chess_insights --cov-report=html
-
-new pytest command w/ module changes:
-pytest -v --cov=chess_insights --cov-report=html
